refactor: split causal/causal_impact.py into 3 modules (#226)
- Extract CausalImpactResult dataclass to causal_impact_results.py (82L)
- Extract reporting methods (results, summary, to_frame, placebo_test) to
  CausalImpactReportingMixin in causal_impact_reporting.py (155L)
- Main causal_impact.py reduced from 648L to 449L (under 500L target)
- All existing import paths preserved via mixin inheritance
- Add 8 structural tests verifying split, backward compat, and line count
- 38/38 tests pass (30 existing + 8 new)
